Unit tests 93%, excluding browsertests
	modified:   src/HTTP.jl
separate _openstream(..) from open for test access
	modified:   src/WebSockets.jl
future improvements
	modified:   test/runtests.jl
	deleted:    test/HTTP.jl
	new file:   test/client_server_test.jl
	new file:   test/client_test.jl
	new file:   test/error_test.jl
	new file:   test/frametest.jl
	new file:   test/handshaketest.jl
